Better solution for avoiding double registering
Previous solution misused bundle contexts and link between factory and original Python module
was lost.
This link is now done again, by relying more on iPOPO.
Drawback is that a factory cannot be used where it is registered, but in real case, it would
be awkward. Docstrings have been updated to highlight this "limitation"