renamed "dcn_beta" to "aerodynamics:vertical_tail:cruise:CnBeta", added "aerodynamics:fuselage:cruise:CnBeta"
dcn_beta was "the difference between fuselageCnBeta and the CnBeta to attain".
And as it is expected to be compensated only by vertical tail, it becomes the CnBeta that is needed from vertical tail.

It looked more logic to have the CnBeta computed for fuselage, then compute VT area in the same time as its CnBeta.