Registering of internal module is now done in one file
BundleLoader().install_packages() is good for installing modules from file system, which will
apply to user modules.

But for internal modules, it will fail if installation is done as archive. Therefore, a more
direct way to register these modules is preferred. It is done in an iPOPO activator to ensure
it won't be done twice

The other part of code that was relying on the install_package() mechanism was the subfolder_provider,
but:
  - it could be used without iPOPO
  - it is actually not used, so it is removed