Fuel computation change
Changed the performances modules so that its natural output (the fuel consumed) is now the only output and separated the mtow update in its own folder. Done for every notebooks. Also favored as much as possible the use of the mtow_loop_and_fuel function and deleted the compute fuel for pure python.
Graphs and paths have not been updated yet !