MissionViewer fixed
Solved the issue there was with the mission viewer that caused it not to update even when changing which variable was to be displayed. Now using and Output widget to solve the issue. Also added the possibility to choose which mission to display when calling display(). If no names are called, the first added mission will be chosen. Raise an error when an invalid name is given.