Refactoring of the generation of configuration and data source file
As a preamble, based on what was available in the documentation, the term "user file" was used as a generic for both configuration files, source data file and input/output file (though not used here).
Both methods are now based on the _generate_user_file which the does almost the same as generate_configuration_file expect it checks what type of user file is supplied (mainly if is among a list of possible one defined in _AVAILABLE_USER_FILE_TYPE). It then calls the method that lists user files of that type and their info based on an if/else statement. Could be improved with a dictionary translating user file type into the associated function (or maybe use eval but it doesn't seem safe).