Bug fix
When taxi-out and takeoff were automatically added (for backward compatibility), output variables for these 2 phases were not created.

Also, in that case, variable prefix could be ignored for input variables, which is also fixed, by using the mission generated "structure", which contains the added phases, instead of the mission definition, which is simply the parsed file (hence without the added phases).