Add RunOnce solver as default
When using the `run_system` metod to perform tests, the default solver of the `BaseCycleGroup` class is used which is `NonlinearBlockGS`. This means that all component tested using this method are ran at least twice (enough for the solver to consider it is done) even if their is no point in doing it. This commit change the default to NonLinearRunOnce